# perf(gatsby-plugin-mdx): prevent babel parse step at sourcing time