… packages/gatsby-plugin-mdx/utils/__tests__/__snapshots__/import-parser.js.snap