The mdx plugin was doing its default parsing step for every time it got called. At sourcing time it's only called to retrieve the import bindings and for this we can be much faster by manually processing the import statements. So that's what this does.

This change __reduces _baseline_ mdx processing by 30%__ by skipping a Babel parse step in one of the bootstrap steps

A very simple flat no-image mdx benchmark cuts down sourcing time in half for this change.

## Benchmark numbers

This are the numbers for a simple, flat, local sourced, zero-image, mdx benchmark where each mdx file is automatically generated and look like this:

```
---
articleNumber: 1
title: "Maiores iusto autem dolorum."
path: 'maiores-iusto-autem-dolorum.'
date: 2018-04-14
---

import { Link } from "gatsby"

<Link to="/">Go Home</Link>

Reiciendis assumenda molestiae a. Ea aperiam aut numquam id. Ratione sint saepe illum est porro asperiores quas ut tempora. Minima sit enim deserunt maxime.
 
Earum iure et explicabo quis qui inventore modi commodi qui. Dolorem ab quo minima neque suscipit voluptas odio deleniti occaecati. Dolore temporibus ut fugiat facilis maiores dignissimos ut eius. Eius similique eos aut magnam quaerat modi. Autem dolor saepe quibusdam reprehenderit error et aut illo.
```

So there's a header, an import, a link, and two paragraphs of random garble (pre-generated, not part of the timings). The benchmark can run with an arbitrary number of pages (`N=1000 yarn bench`) so that's what we do for 1k, 2k, 4k, 8k, and 16k. By 2x-ing every step we can detect logaritmic perf problems easier. While the baseline Gatsby pipeline holds for a million pages easily, for mdx that ceiling is a lot lower right now (that's what I'm trying to improve :p )

The benchmark runs on a dedicated Intel NUC, which is a headless linux setup, specifically to benchmark, so numbers ought to be fairly consistent relative to other runs.

## Data

Master is currently roughly on gatsby-plugin-mdx@1.2.19 and gatsby@2.23.12 ( + whatever is unpublished).

The raw data for both runs can be found in two gists:
- master: https://gist.github.com/pvdz/10eb6b92f58a32233c10dd9f8810d7b3
- this PR: https://gist.github.com/pvdz/b1556a3b51476af05500d4f0ba433c91

----

- 1000 pages:
  - master: 
    - sourcing: 30s
    - bootstrap: 33s
    - finished: 59s
  - PR:
    - sourcing: 14s
    - bootstrap: 17s
    - finished: 45s
  - delta: -14s
  - e2e page speed 16p/s -> 22p/s
- 2000 pages:
  - master
    - sourcing: 55s
    - bootstrap: 59s
    - finished: 104s
  - PR:
    - sourcing: 25s
    - bootstrap: 29s
    - finished: 76s
  - delta: -28s
  - e2e page speed 19p/s -> 26p/s
- 4000 pages:
  - master
    - sourcing: 107s
    - bootstrap: 112s
    - finished: 193s
  - PR:
    - sourcing: 48s
    - bootstrap: 52s
    - finished: 137s
  - delta: -56s
  - e2e page speed 20p/s -> 29p/s
- 8000 pages:
  - master
    - sourcing: 221s
    - bootstrap: 227s
    - finished: 388s
  - PR:
    - sourcing: 104s
    - bootstrap: 110s
    - finished: 281s
  - delta: -107s
  - e2e page speed 20p/s -> 28p/s
- 16000 pages:
  - master
    - sourcing: 464s
    - bootstrap: 474s
    - finished: 819s
  - PR:
    - sourcing: 218s
    - bootstrap: 226s
    - finished: 572s
  - delta: -247s
  - e2e page speed 19p/s -> 27p/s (I mean, 27.97 ...)