@pvdz 2020-07-07
Confirmed a canary to build (https://github.com/gatsbyjs/gatsby/pull/25569/checks?check_run_id=845921135)